Next Button in DraftVideoListPage: Added a Next button in the DraftVideoListPage to allow users to proceed to the EditDraftPage for editing their draft content.
EditDraftPage: Created the EditDraftPage, where users can edit the content of their draft. This page initially fetches data from the database to populate the fields.

Data Fetch from Database: Implemented the retrieval of draft data from the database and used this data to populate the fields in the EditDraftPage.

Data Update: Users can update the data in the EditDraftPage fields, and when they click the 'Update Draft' button, the updated data is saved back to the local database.

ImagePopUpWithOk Feature: Introduced the ImagePopUpWithOk feature, enabling the creation of single-button dialog boxes with dynamic data and images for user interaction and feed